Added customizable palettes
I added a few default palettes to the GBPalettes class and an array
constructor so that a palette can be made from any array of four colors.

Note that using these would not simply affect a palette swap on the
normal result… Because the colors that the Gameboyifier compares per
pixel are totally different, the image becomes totally different too!