Added palette labels
I added a "label" field to the GBPalette class, with a getter method. If you make a palette using one of the static final defaults, it should compare the addresses of the array you set with the possible arrays and decide on a label. (There has to be an easier way to do that... I mean, not only is it right in the variable name, but since we have the address already don't we already know where to look?)