Brought code out of the constructor
The code for getting images and USING a Gameboyifier really doesn’t
belong in a constructor… As such, most of it has been moved to main.
The code for reading and writing images should maybe also be moved at a
future date, for consistency. That or more of the main method should be
encapsulated. I can’t decide.